refactor(visual): 重构视觉编辑功能中的类型定义和变量命名
- 将AI工作区面板中的预设操作从FIX_OVERFLOW/BALANCE_COLUMNS改为RESIZE_ELEMENT/MOVE_ELEMENT
- 为建筑图标注册服务添加Callable类型注解和ArchitecturalIconMatch导入
- 更新资产匹配服务中的返回值转换确保布尔类型一致性
- 修正幻灯片交付状态计算中的变量命名避免覆盖问题
- 优化导出策略服务中的循环变量命名提高代码可读性
- 为导出往返测试服务添加渲染器快照类型注解并改进MSE计算逻辑
- 更新UI页面中的队列指标

+++ archium/application/visual/placeholder_binding_matcher.py
@@ -44,12 +44,12 @@ def geometry_similarity(
         bcx, bcy = bx + bw / 2, by + bh / 2
         dist = ((acx - bcx) ** 2 + (acy - bcy) ** 2) ** 0.5
         diag = max((aw**2 + ah**2) ** 0.5, (bw**2 + bh**2) ** 0.5, 0.01)
-        return max(0.0, 1.0 - dist / diag) * 0.35
+        return float(max(0.0, 1.0 - dist / diag) * 0.35)
     inter = (right - left) * (bottom - top)
     union = aw * ah + bw * bh - inter
     if union <= 0:
         return 0.0
-    return inter / union
+    return float(inter / union)
 
 
 def score_placeholder_candidate(
